move functions from cahelper to other files
